Pushing local chances; at this point in time the code base has not been touched in almost a year so I'm not exactly sure what changes were made here or whether or not anything is broken

+++ sys/kern/file.c
@@ -264,10 +264,10 @@ fop_write(struct file *fp, const char *buf, size_t nbyte)
     struct fops *ops = fp->ops;
 
     if (ops->write) {
-        int written = ops->write(fp, buf, nbyte);
+        ssize_t written = ops->write(fp, buf, nbyte);
 
         if (written > 0) {
-            fp->position += written;
+            fp->position += (off_t)written;
         }
 
         return written;
